lua: Restructure lua modules.
