+++ Source/smvluacore/constants.lua
@@ -1,0 +1,7 @@
+-- predefined axes
+X = 1
+Y = 2
+Z = 3
+
+orthogonal = 1
+perspective = 0
