+++ Source/smvluacore/unload.lua
@@ -1,0 +1,12 @@
+--- @module unload
+local unload = {}
+
+function unload.all()
+    unloadall()
+end
+
+function unload.tour()
+    unloadtour()
+end
+
+return unload
