+++ Source/smvluacore/window.lua
@@ -1,0 +1,7 @@
+--- @module window
+local window = {}
+window.size = function(width, height)
+    setwindowsize(width, height)
+end
+
+return window
